Fix backend-postmaster-autovacuum audit FAIL to a faithful PASS
Address the five findings in audits/backend-postmaster-autovacuum.md:

- M1: extract_autovac_opts was MISSING (dissolved into the catalog seams,
  which returned pre-extracted AutoVacOpts). The scan/recheck carriers now
  carry the raw extractRelOptions result (Option<StdRdOptions>); the
  .autovacuum projection + relkind assert is ported in-crate as
  core::extract_autovac_opts and called at all three C call sites
  (do_autovacuum two passes, table_recheck_autovac).

- D1: do_autovacuum now reproduces the per-table PG_TRY/PG_CATCH error
  isolation. autovacuum_do_vac_analyze is matched (not `?`-propagated): on
  Ok it clears QueryCancelPending; on Err it builds autovacuum's own
  errcontext("automatic {vacuum,analyze} of table ...") line in-crate and
  hands the adorned error to a foreign emit-report-and-restart seam, then
  continues with the next table. did_vacuum/switch-to-AutovacMemCxt match C.

- D2: the orphan-temp recheck/drop control flow is ported in-crate over
  fine-grained leaf seams (ConditionalLockRelationOid / re-fetch class row /
  ConditionalLockDatabaseObject / performDeletion / snapshot+xact), with the
  recheck predicates (relkind/relpersistence, TEMP_NAMESPACE_IDLE) and the
  LOG("dropping orphan temp table") decision in-crate, replacing the single
  drop_orphan_temp_table seam.

- D3: autovac_report_workitem now bounds the appended suffix to
  MAX_AUTOVAC_ACTIV_LEN like C's snprintf(activity+len, MAX-len, ...) via a
  shared snprintf_append helper (also used by autovac_report_activity).

- D4: rebuild_database_list truncates millis_increment to i32 before the
  <= MIN_AUTOVAC_SLEEPTIME compare and on the *1.1 store, mirroring C's
  `int millis_increment`.

New carrier OrphanClassRow + StdRdOptions-typed relopts fields in
types-autovacuum; new fine-grained seams in the ext-seams crate. Added unit
tests for extract_autovac_opts and snprintf_append. Gate: cargo check
--workspace + cargo test --workspace both green.

+++ crates/backend-postmaster-autovacuum/src/launcher.rs
@@ -361,11 +361,19 @@ pub fn rebuild_database_list(newdb: Oid) -> PgResult<()> {
         /* sort the array by score */
         dbary.sort_by(|a, b| db_comparator(a, b));
 
-        /* Determine the time interval between databases in the schedule. */
-        let mut millis_increment: f64 =
-            1000.0 * core::autovacuum_naptime() as f64 / nelems as f64;
-        if millis_increment <= MIN_AUTOVAC_SLEEPTIME {
-            millis_increment = MIN_AUTOVAC_SLEEPTIME * 1.1;
+        /*
+         * Determine the time interval between databases in the schedule.
+         *
+         * C declares `int millis_increment`, so the float quotient is
+         * truncated to an int *before* the `<= MIN_AUTOVAC_SLEEPTIME` compare
+         * (the int is promoted back to double for that compare), and the
+         * `MIN_AUTOVAC_SLEEPTIME * 1.1` (= 110.0) is likewise truncated to 110
+         * on store.  Mirror that with an explicit `as i32`.
+         */
+        let mut millis_increment: i32 =
+            (1000.0 * core::autovacuum_naptime() as f64 / nelems as f64) as i32;
+        if (millis_increment as f64) <= MIN_AUTOVAC_SLEEPTIME {
+            millis_increment = (MIN_AUTOVAC_SLEEPTIME * 1.1) as i32;
         }
 
         let mut current_time = seam::get_current_timestamp::call();
